move `sendChunkedBody` function. remove unnecessary usage of `this`
